locations as button, as it is slow for pokemon with many locations
